moved maze editor to proper package
