+++ src/grp15/Main.java
@@ -1,6 +1,6 @@
 package grp15;
 
-import grp15.object.MazeEditor;
+import grp15.simulator.MazeEditor;
 import grp15.object.Robot;
 
 import java.awt.BorderLayout;
@@ -9,12 +9,8 @@ import java.awt.Dimension;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 
-import javax.swing.DefaultComboBoxModel;
 import javax.swing.JButton;
-import javax.swing.JComboBox;
 import javax.swing.JFrame;
-import javax.swing.JLabel;
-import javax.swing.JOptionPane;
 import javax.swing.JPanel;
